services/headysoul/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
sentence-transformers==2.3.1
torch==2.1.1
transformers==4.36.2
numpy==1.24.3
faiss-cpu==1.7.4
python-dotenv==1.0.0
pyngrok==7.0.0
huggingface_hub==0.19.4
